+++ src/main/java/com/technophobia/substeps/runner/runtime/StepClassFilter.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright Technophobia Ltd 2012
+ *     Copyright Technophobia Ltd 2012
  *
  *   This file is part of Substeps.
  *
@@ -18,6 +18,7 @@
  */
 package com.technophobia.substeps.runner.runtime;
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 
 import com.google.common.base.Predicate;
@@ -26,8 +27,35 @@ import com.technophobia.substeps.model.SubSteps;
 public class StepClassFilter implements Predicate<Class<?>> {
 
     public boolean apply(final Class<?> clazz) {
+        if (hasAnnotation(clazz, SubSteps.AdditionalStepImplementations.class)) {
+            return true;
+        }
         for (final Method method : clazz.getMethods()) {
-            if (method.isAnnotationPresent(SubSteps.Step.class)) {
+            if (hasAnnotation(method, SubSteps.Step.class)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+
+    private boolean hasAnnotation(final Method method, final Class<?> annotationClass) {
+        return hasAnnotation(method.getAnnotations(), annotationClass);
+    }
+
+
+    private boolean hasAnnotation(final Class<?> clazz, final Class<?> annotationClass) {
+        return hasAnnotation(clazz.getAnnotations(), annotationClass);
+    }
+
+
+    private boolean hasAnnotation(final Annotation[] annotations, final Class<?> annotationClass) {
+        // compare the class names rather than the classes themselves, as its
+        // entirely possible that the class was created in a different
+        // classloader to
+        // the current classs (e.g. if this is the eclipse plugin)
+        for (final Annotation annotation : annotations) {
+            if (annotation.annotationType().getName().equals(annotationClass.getName())) {
                 return true;
             }
         }