+++ src/main/java/com/technophobia/substeps/runner/runtime/StepClassLocator.java
@@ -23,4 +23,9 @@ public class StepClassLocator extends PredicatedClassLocator {
     public StepClassLocator(final String path) {
         super(new StepClassFilter(), new ClassLoadingFunction(path));
     }
+
+
+    public StepClassLocator(final String path, final ClassLoader classLoader) {
+        super(new StepClassFilter(), new ClassLoadingFunction(classLoader, path));
+    }
 }